requirements.txt
Django
djangorestframework
django-filter
django-cors-headers
psycopg2-binary
djangorestframework-simplejwt
requests
djoser
Pillow